requirements.txt
streamlit
pandas
plotly
st-gsheets-connection
google-generativeai